Refined error message for failed sortition pool status monitoring
In 99% cases this error is logged when the sortition pool status
monitoring failed because the operator was removed from the sortition
pool because it does no longer have enough unbonded value or stake.

We do not change the logic in any way here but just improve the error
message to at least give the operator some context on what they should
be looking at.