Make signer snapshot just after generating it
Signer snapshot should me made just after
generating it and before the public key
is submitted on chain. With that behavior,
we can safely store the signer in the
persistence layer and have the possibility
to recover in case something goes bad
after the public key is registered on chain.