Register applications for FullyBackedBondedECDSAKeepFactory
For keep members banning we need to have a list of all applications that
got registered to track their sortition pools. In this commit we updated
the function for sortition pool to additionally register application
address in an array. Later we can use the array to loop over application
to sortition pool mapping.
We also had to update createSortitionPool function from `external` to
`public`.