Ban keep members from FullyBackedECDSAKeepFactory
Exposed a function to ban in all sortition pools members of keep calling
the function. It will loop over all created sortition pools associtated
with this factory and ban all members of a keep. A keep is determined as
a caller of the function. The caller can only be a keep created by the
factory.