Ban keep members from FullyBackedECDSAKeepFactory
Exposed a function to ban in a sortition pool members of keep calling
the function.

A keep is determined as a caller of the function. The caller can only
be a keep created by the factory.

This function requires linking keep with an application address for
which keep was created. We introduced a mapping filled on a keep
creation.