Minimum keep count to obtain ECDSA reward set to 1000
To make unit tests possible, there was a need to trick keep creation in
the stub contract. Instead of cloning real contract, we just append
generated addresses. Also, we are not able to test all intervals given
the execution time of tests so we cover the first three and rely on
Reward contract unit tests to do the rest.