Hex up `signing sign-digest` signature output and include pubkey
In particular, the signature is now output as a single 128-character
string consisting of two 0-padded 64-character hex representations of
the signature R and S values, in order, and the public key is outputted
as a concatenated X/Y coordinate pair. These two outputs can be used
directly with the new bin/tbtc.js support for constructing signed
transactions.