Cap the maximum reward per ECDSA keep operator per interval
Set a limit on the amount of KEEP a single ECDSA keep operator can receive
in a single reward interval.
Use the withdrawal pattern to track allocated and withdrawn rewards.
Rewards exceeding the cap will be returned to the unallocated pool.

Because the reward payouts are changed to the withdrawal pattern,
the existing ECDSA rewards tests should not work.
For some reason this is not the case and the tests pass.