Add documentation and fix ordering issue
When a reward allocation to an operator exceeds the reward cap,
the reward allocation is to be reduced and the excess unallocated.
Two lines were transposed leading to the unallocated amount always being zero.
Fix this by reordering the lines.

Also, add documentation explaining the reward caps, allocations and withdrawals.