Fixed assertKeepBalanceOfBeneficiaries
This function was not asserting anything because of a wrong construction
of the loop. We were missing `.length` invocation there.

This commit fixes the assertion as well as outdated tests that were
passing because of the broken assertion although they should not (not
minimum keep quota in the interval met).