Added top up function to Fully Backed Bonding contract
We defined a `topUp` function in `FullyBackedBonding` contract. This
function should be used when adding new value for bonding. We emitt
an additional event to clearly find value that has been deposited. The
`deposit` function defined in the abstract contract can be used also by
applications that return part of seized value to the contract. This
made calculation of totally deposited value much more difficult.