ECDSA staker rewards cap set to 3M keep tokens per interval
Individual beneficiary can not receive more than 3M KEEP tokens in a
single interval. This amount is about 20% of the second interval full
allocation.

I have made beneficiaryRewardCap public constant. It's a different
situation to the rest of constants defined in ECDSARewards as they are
only used as a constructor parameters for the parent contract.

Added test to make sure this value is set at 3M * 10^18.