Adjust ECDSA keep rewards by lifespan
Reward includes a constant fraction that is paid to all happily closed keeps,
and a variable fraction which is paid based on the lifespan of the keep.
Keep lifespans are estimated by having users report closed keeps.
If the closure of a keep has not been reported,
assume it lasted long enough to be eligible for maximum reward.