Report open instead of closed keeps, and include bond amounts
By reporting that a given keep is still open instead of waiting for it to close,
we are able to query the value of the bonds for the keep.
This allows the adjustment of reward values
based on the bonded amount multiplied by the duration it has been locked up.
This is a more precise and preferable method.