Increase fee by a step from the two latest events
Deposit requires the fee to be bumped by the constant value equal the
initial redemption fee. To obtain the initial fee we looked for the very
first event, we also need the latest event. It worked fine until there
were multiple fee increases and the number of blocks to reach the first
event exceeded 10k that is a default limitation for get past events. In
such situation the first event that was returned wasn't actually the
very first event registered on chain for the deposit.
To overcome this issue we decided to calculate the initial fee as a
delta between fees of the two latest events, as the step value is constant
for each fee increase.