Updated signatures to ethereum's Text Sign format
Ethereum is hasing a message for signing with a dedicated hashing
function that is adding more information to a signed message:
keccak256("\x19Ethereum Signed Message:\n"${message length}${message})

In this commit we are using ethereum's hashing function to obtain digest
to sign.

We also extracted parts of the functions to be able to test them.

Having these changes signatures should be verifiable on mycrypto.com
site.