Detect fraudulent operators
We want to check if an operator commited a fraud. To find that we look
for `TokensSlahed` events emitted by `TokenStaking` contract.

After fetching the events we need to filter out the ones that originated
in Keep ECDSA.

A fraud can be reported by calling BondedECDSAKeep contract directly or
by calling related Deposit contract.

For each discovered event we get a contract which was called in a
transaction that emitted the event. If the contract is BondedECDSAKeep
or Deposit we mark an operator as fraudulent.