Replacing receiveApproval function with a seperate allocate function
Allocate amount of KEEP for the given merkle root. Calling
receiveApproval with the onlyOwner modifier results in error because
caller is msg.sender which is the KEEP Token. But what we need is the
caller (owner) who is the owner of Distributor contract as well.
We need to replace it with a seperate function where the owner can be
set explicitly.