Deploy ECDSARewards
The ECDSARewards should be deployed to network. The only reason why
rewards contracts are not yet on Ropsten is we were not doing an RC cut
for rewards contracts. The need to deploy the ECDSARewards also occurred
while working on the ecdsa rewards on the KEEP token dashboard side.

Here we also:
* added a fn in `lcl-provision-external-contracts` which fetches the
KEEP token adress and save in `external-contracts` file. The KEEP token
address is required to deploy the ECDSARewards,
* added a stub function which allocates fake rewards for a given
operator in the given interval. This function helps testing rewards in
KEEP token dsahboard,
* added a script which allocates fake rewards via ECDSARewardsStub
contract.