Updated keep-core contracts dependency to 1.5.0-pre.1
The new version contains `receiveApproval` function on `PhasedEscrow`.
It was added in keep-core commit 5dfd8565054288588f73ee67d238ec93f0cbcfd0
that was added after solidity/v1.4.1 release or 1.5.0.pre.0.

We need `receiveApproval` so that ECDSARewardsDistributorEscrow which is
a variation of PhasedEscrow can receive tokens via approveAndCall from
another PhasedEscrow with StakerRewardsBeneficiary.