Setup tenderly on context initialization
We wan't to initialize tenderly with API access key on context
initialization. Tenderly is optional, as not every user has access to
it. If the API Key is provided we will initialize Tenderly. If the key
is not provided Tenderly will be missing and we won't call it but depend
on cached data.