Authorization conditions checks
We need to verify authorizations at interval start and deauthorizations
made during the interval.

Unfortunatelly deauthorization doesn't produce any event, so we need to
fetch method calls from the chain. To do that we use Tenderly as instead
we would need to go block by block looking for transactions. Tenderly
requires API access key, not every user will have it, so we will run the
script and cache the data.

Authorizations at interval start are checked by running the functions
with block parameter.