Stake and unbonded value checks at interval start
We wan to check additional conditions at interval start:
1. Check if operator staked minimum value, if not they won't get any
rewards.
2. Check if operator had minimum unbonded value available. If so verify
if operator was registered in the pool. If operator had minimum unbonded
value and was not registered in the pool the won't get any rewards. If
operator didn't have minimum unbonded value available they still get a
chance for rewards as they can have all the ether locked in bonds
already.