Reorganizing staker rewards structure to avoid dependeny errors
We use node v14.3.0 with --experimental flag to run rewards.js script
with "type:modules" in package.json. This causes failures when running a
script to generate a merkle objects. It works with node v11.15.0 but not
with v14.3.0 This commit is a little workaround to make scripts work together
nicely without major refactoring.