Transferring KEEP rewards in notifyRewardAmount
Instead of adding approveAndCall function, it is cleaner to add a
'transfer' function in notifyRewardAmount which has to be called anyway
by the reward distributor. Reward distributor in our case is escrow's
beneficiary which is going to be set by PhasedEscrow contract.

In tests, the role of PhasedEscrow beneficiary takes rewardDistribution
account. We need to transfer funds from KEEP token owner to rewardDistribution
address first, before we can call notifyRewardAmount().