Ignore sub-project node_modules for eslint
Eslint ignores node_modules in the project root but it was going through
node_modules of staker-rewards sub-projects: merkle-distributor
and distributor. This was leading to JavaScript heap out of memory
failures during linting. To see that eslint goes through those
directories, it is enough to add --debug flat to lint:fix:js execution
of eslint.

The solution is to ignore
/merkle-distributor/node_modules
/distributor/node_modules
in .eslintignore what has been done here.