Log intervals when switching to partial event pulls
When getPastEvents call fails, staker rewards script switches to partial
event pulls, traversing in 2000 blocks intervals. This takes time,
was not reflected in the execution log, and was confusing - it looked
like if the script hanged completely.

With this commit, partial event pulls informs about the progress by
logging information about what range is currently fetched, how many
events it contains, and what's the total number of events so far to the
console.