Undelegating operators in staker rewards script
If an operator doesn't meet the sortition pool
requirement at the interval start, it may be the
case that they are currently undelegating. In such
a case, that operator should not be punished
for requirements violation. Instead of that,
only the `ETH_bonded` value from their still
active keeps should be count to their
`ETH_locked` value. The `ETH_unbonded` value
is ignored.