Deduplicator for key-gen and signature request events
After the changes in Ethereum subscription API, it is possible some
quite an old event will be emitted. We need to have a better event
filtering logic in place that will filter events that are currently
being processed but also those events that were already processed and
their processing finished.

Here, such a logic was implemented for key-gen and request signature
events for a keep. For now, it is just a standalone code not attached
anywhere.