Abstracted tracker code for simple keep events
Abstracted the code from requestedSignersTrack to a separate struct to
reuse it for other cases, such as closeKeepTrack or archiveKeepTrack.

All those cases are events happening once in the entire keep history and
they do not require any special tracking with additional fields allowing
to distinguish them.

closeKeepTrack and archiveKeepTrack will soon be used in event
deduplicator to deduplicate keep close and archive events.