Deduplicator functions for close and terminate keep events
Close and terminate keep events are the last two events that should be
deduplicated. Not filtering out a duplicate is not as dangerous as in
case of key-gen or signing events but lets to avoid errors in logs about
keep not existing in the storage.

Additionally, did some renames to better reflect what is actually
happening in the deduplicator. Specifically, deduplicator is not
generating a key or signing. It is just getting notified about the
client wanting to proceed and helps the client to decide if it should
proceed or not.