Generic retry function implementation with an exponential backoff
client.go and node.go implements retries for key generation and
redemption signatures. Retry logic is mixed with the core logic managing
the flow. This entire logic reached its critical mass and it is now very
hard to add there any call that can produce an error.

In one of the next commits, such a call to event deduplicator will be
added. To make this task easier, here there is an implementation of a
generic retry function that can be reused in several places.