Improved logging in OnBondedECDSAKeepCreated handler
We now log the block number at which the event was emitted as well as
information whether the client is a member in this keep.

After event handler refactoring, this event can be received by the
client code multiple times and as a result, logged multiple times.
The improved logging will make it clear for operators that
this is an event from the past and they should not be worried about it -
either it was already handled, or the client is not a signing member so
the event can be safely skipped.