Confirm awaiting signature with a timeout
Instead of executing awaiting signature check in the redemption
signature event deduplicator only one time, we now do it multiple times
with 30 seconds timeout. One check could not be enough in case of a
minor chain reorg or in case when chain clients are not perfectly in
sync. Specifically, this will address Infura problems where it usually
takes a couple of seconds for nodes behind a load balancer to sync about
the chain state.