Confirm deposit state before starting monitoring
After Ethereum event subscription refactoring, it may happen that the
same event is emitted multiple times. A duplicate can be emitted right
after the original event or a long time after the original event has
been emitted.

The original version of tBTC deposit monitoring code - without the
modification here - was mostly safe. There is 24 hours monitoring period
eliminating duplicate events within this time frame.

Just to be absolutely safe, in case 24 hours time period is exceeded for
a duplicate, we are adding initial state check. Before monitoring starts
on start function signal, we check the initial deposit state whether
it's an expected one. If not, we do not start the monitoring.