Accept string in deduplicator interface as keep address
common.Address is ethereum type and is not chain-agnostic. To do not
propagate ethereum-specific types any further in the codebase we update
deduplicator's public interface into string-based approach.

There is a bunch of transformations back and forward we have to do now
but they should all be gone once we switch to string-based approach
everywhere.