Changed the order of cache checks in shouldMonitorDeposit
The previous approach assumed that if the given deposit address is
cached in monitoredDepositsCache, shouldMonitorDeposit returns true
withoug executing any further checks.

Together with acquireMonitoringLock and releaseMonitoringLock it worked
fine for most of the cases but not for all of them. Specifically, for
the case when the deposit has already transitioned to the next state, we
could run into a situation when monitoring lock has already been
released and shouldMonitorDeposit returns true just because the client
has a member in that deposit, without executing the initial state check.

The order of cache checks has been changed in this commit to address it.

Also, renamed monitoredDepositsCache to memberDepositsCache, and
notMonitoredDepositsCache to notMemberDepositsCache to state it clearly
those caches hold information whether the client has or has not a member
in the given deposit.

The new approach is:
1. Check if the given deposit is in notMonitoredDepositsCache. If so, we
are sure the client is not interested in monitoring this deposit no
matter what - it has no member in it.

2. Check initial state of the deposit to do not start monitoring for
duplicate events from the past. If initial state is not as expected, the
client is not interested in monitoring this deposit as the event
triggering the monitoring is something from the past that does not match
the current chain state.

3. Check if the given deposit is in memberDepositCache. If so, start the
monitoring. This check allows to do not repeate getSignerIndex call. If
the deposit is not in memberDepositCache, execute getSignerIndex call,
check members, update caches and return true or false depending if the
client has a member in the deposit.