Use Blockstream fallback as node connection instead of per action
The section on address derivation explicitly called out attempting to
use the Blockstream API as a fallback. Instead, it now focuses on using
the optional node connection described in the Electrs section, and that
section defines the Blockstream API as a fallback node connection. It
also allows disabling that connection entirely if preferred.