Better use btcd in transaction and sighash construction
The signing process now starts by constructing an unsigned transaction,
then using btcd functionality to derive its sighash. Once the sighash is
signed, the final transaction construction consists simply of attaching
the witness signature to the existing unsigned transaction.

This makes most of the code significantly easier to follow, and reduces
the number of parameters taken in each function, which happens to make
the code easier to follow as well. It also fixes the fee handling, which
before would have resulted in an incorrect signature for the transaction
as fees were adjusted post-signing.