Updated .dockerignore file
Updated file with files and directories that we don't want to copy to
Docker building context.

We can ignore all hidden directories and files, infrastructure, scripts
and staker-rewards directories, more files and adocs as they are not
used by the Docker image.

We also ignored content of solidity directory with an exclusion for
contracts and NPM package configuration that is needed for Go code
generation.

These changes reduced data sent to the docker building context on local
machine from ~1.5 GB to ~2.5 MB.