Add BroadcastRecoveryAddress
This is a proof-of-concept function which almost certainly lives in the wrong file. Here, we set oureslves up to be able to test and verify that we can broadcast and sychronize btc addresses in the case of a liquidation event before we spend too much time on deciding the best place for this code to live.