Removed rfc-18 branch filter for CircleCI job execution
Initally we added ignore filter for branches with rfc-18 prefix which
sounds like a good idea taking into account that we want to run GitHub
Actions for such branches, so duplication is not necessary. It turned
out that we require CircleCI job to be executed as a check before
merging to master. That's why we removed the filter and will execute the
job anyway. It shouldn't be much of a problem as the duplication will
take place only for workflows on rfc-18 branches as we filter them in
GiHub Actions configuration.