Create GH Actions workflow for build, test nd lint of Solidity contracts
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvment and make
the processes faster and less error prone. A part of this task is
migration from CircleCI jobs to GitHub Actions. This commit creates a
GitHub Action workfow for building, testing and linting of Solidity
contracts.

For now, execution of the workflow has been limited only to branches
which name starts with 'rfc-18/'. Although introduction of the workflow
potentially allows us to exclude branches 'rfc-18/*' from
build_and_test_solidity step in CircleCI, this has not been done, as
passing of this step is a requirement for merging of the pull requests.