Slightly simplify/shorten witness construction
By using wire.TxWitness instead of [][]byte, we gain some clarity in the
code's intent. SigHashAll is a one-byte value so it doesn't need to go
through endianness handling in a byte-buffer. And finally, the
compressed bytes for the public key can be achieved in-place, vs
referencing a function that wasn't defined anywhere in the RFC.