Remove direct imports of `go-ethereum/crypto` pkg
All references to the `go-ethereum/crypto` should
not be present when building `keep-ecdsa` with
`celo` build tag. This is because this package
conflicts with the `celo-blockchain/crypto` as
they both use the same native C code under the
hood. This conflicts cause C linker error during
`go build` due to duplicated symbol.