Rename pkg/chain package declaration to `chain`
The pkg/chain package was named `package eth` throughout its files due
to a historical rename that wasn't fully executed. The package is now
correctly named `chain`, and all references to it have been revised to
correctly refer to it as `chain` instead of `eth`. This is in
preparation to remove from the interface the Ethereum-specific types and
make the package a true cross-chain abstraction.