Pull balance monitoring and tBTC setup into ethereum.Connect
This removes the requirement to expose internal chain details to the
commands that boot up the Ethereum-flavored client. In fact, the
tBTC setup should be accessible from the command setup, but the current
interface doesn't yet allow that. Pulling tBTC setup back up to the top
level will be done after the relevant interface refactor is complete.