Make BondedECDSAKeepHandle stand alone
The chain.BondedECDSAKeep interface, which was embedded in the
chain.Handle interface, is now the chain.BondedECDSAKeepHandle
interface, which stands as a distinct interface providing interactions
to a single discrete underlying BondedECDSAKeep. BondedECDSAKeepFactory
currently provides two methods to access it, GetKeepAtIndex (repurposed
to return a BondedECDSAKeepHandle instead of a common.Address), and the
new GetKeepWithID (which takes a common.Address and returns a
BondedECDSAKeepHandle). Since BondedECDSAKeepFactory remains embedded
in chain.Handle, these are callable directly on a chain.Handle.

In the process, all methods that were on BondedECDSAKeep no longer take
a keep address as a parameter. As well, many of the functions in
client.go, node.go, and tbtc.go are now implemented in terms of
chain.BondedECDSAKeepHandles instead of common.Addresses. For logging
purposes, chain.BondedECDSAKeepHandle has an ID() method that returns
the underlying keep's common.Address.