Multi-chain Dockerfile
Here we refactor the Dockerfile in order to
support chains other than Ethereum. This is
achieved by introducing a new CHAIN argument
which is used to generate the right set of
contract bindings and is included on the build
tag list. By the way, this change also gets rid
of the go-ethereum's abigen installation. This
is no longer needed due to changes introduced
by PR #481. Last but not least, several
adjustments have been made to make the Dockerfile
work with Celo. Specifically, the `musl` build tag
is now passed while generating and building
the `keep-ecdsa` client with Celo on Alpine
platform.