Drop chain interface TODOs that are no longer relevant
Originally the intent was for on-chain submissions to return a promise
that would resolve once the on-chain submission was considered final.
Since, it has seemed like finality monitoring, because it may trigger
retries, is best accommodated at a higher level than the return value of
the submission.

Submit calls will continue to return an error if the submission fails
due to connectivity issues.