Move publishing of Go client to GH Actions
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvment and make
the processes faster and less error prone. A part of this task is
migration from CircleCI jobs to GitHub Actions. This commit enhances the
GitHub Action `client.yml` workflow with step that publishes built
docker image to `keep-test` Google Container Registry.