Add `migrate-and-publish` job for Celo
Added a `migrate-and-publish-celo` job which
performs contract migration on Celo chain.
For now it publishes artifacts to GCS bucket
only. Publishing to Tenderly and NPM should
be added later.

By the way, adjusted the former
`circleci-provision-external-contracts.sh` script
to be more CI-platform-agnostic and use it from
within the `migrate-and-publish-celo` action
to provision external contract artifacts.