Increased default max gas price to 1000 Gwei
On 2021-02-23 we observed the average gas price on mainnet went up to
even 625 Gwei. Some operators were required to update their config files
as the default value of 500 Gwei was no longer enough.

Given that we observe high gas prices quite recently, this commit bumps
up the default max gas price to 1000 Gwei.

It does not mean it is the price every transaction will be submitted
with. This is the _maximum_ the client can bump up the gas price to if
the transaction was not mined in the specific time frame.