Move building and publishing of Solidity initcontainer to GH Actions
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvement and
make the processes faster and less error prone. A part of this task is
migration from CircleCI jobs to GitHub Actions. This commit enhances
the Solidity workflow with a job building Solidity initcontainer and
publishing it to the Google Container Registry.

The `build-and-publish-initcontainer` job is configured to execute
after successfull execution of `migrate-and-publish-ethereum`. As
`build-and-publish-initcontainer` requires artifacts from
`migrate-and-publish-ethereum`, they are passed between jobs using
`actions/upload-artifact` and `action/download-artifact` actions.
The artifacts will be also availible for manual download from workflow
run page in GitHub GUI (default storage period is 90 days). If stored
artifacts turn out to take too much space, the retention period can be
changed globally in project's settings, or locally in
`actions/upload-artifact` configuration (`retention-days` parameter).