Improved firewall caching assertions
From now on, we do check the number of calls executed to the chain and
we do confirm if some information was picked from the cache or not.
Previously, tests were not actually covering it - the only thing they
were covering was whether the second or third call returns the same
result.