Cache keep index - keep ID mapping between firewall checks
There are ~8 nodes in the network right now which undelegated their
stake. Given that they do not have a minimum stake now, we execute an
additional check to see if the node is a member of at least one active
keep.

All those nodes try to connect to bootstrap at the same time and
bootstrap is executing the additional check for all of them at the same
time.

This check is expensive and it may happen that it can not complete in
less than 30 seconds (concurrency limits, compute unit per second limit
on Ethereum client) and libp2p will retry. Libp2p retries firewall
check, in this case, every 30 seconds and we may end up stacking
firewall checks for the same address - the previous check can still be
in progress while a new check for the same address will be executing.

Increasing or completely removing concurrency limit and/or switching to
Ethereum client not limiting compute units per second helps but we need
to make sure the firewall is as least expensive as possible.

With this change, we add caching to [keep index, keep on-chain ID]
mapping that never changes on the chain. This way, even if two or more
firewall checks are executed at the same time, most requests will be
serve from the cache. It will help serve those requests faster and with
less calls to Ethereum client.