Cleanup afer moving to GH Actions
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvement and
make the processes faster and less error prone. A part of this task is
migration from CircleCI jobs to GitHub Actions.

Now that we have all CircleCI jobs moved to GH Actions (and verified),
we can implement some final changes removing obsolete CircleCI configs
and cleaning up pieces of code used for workflows testing. Here is a
list of changes applied:
- removed unnecessary comments
- cleaned `config.yml` (CircleCI jobs are no longer needed)
- removed memory monitoring step in client workflows
- modified conditions for workflow triggering and artifacts publication
  (according to the decision that we should migrate contracts and
  publish artifacts to registries after accepting PRs i.e. merging to
  master)
- removed `-wip` suffix from published images
- removed `--dry-run` flag when publishing to npm registry
- removed `circleci-migrate-contracts.sh` which is no longer used
  (`ci-provision-external-contracts.sh` is used instead)