Use env property to get TBTCTSystem address in init script
Adding a dependecy to TBTCSystem contract to obtain it's address in the
initcontainer gives an additional cyclic dependency to tbtc which
already depends on keep-ecdsa. We used to handle this case by defining
the TBTCSystem address as environment variable for the pod
(SANCTIONED_APPLICATIONS). With this change we go back to this solution.

We will revisit pulling the address from a contract artifact when
working on RFC-18 modules integration.