First pass at implementing derivation index persistence
The core idea is that we create file-based storage where each extended public key gets its own folder, and each index gets its own file within that folder. The filename would be the index number and the file contents would be the derived wallet address.

This allows a user to explore their filesystem and parse it for the wallet addresses they've already generated, as well as see which indexes they have already used.