Update version of keep-ecdsa package
To avoid mixup of versions of migrated contracts test and dev
environments must operate on different versions of `keep-ecdsa`
package. Packages with `-rc` suffix will be used on test env (Ropsten).
Packages with `-pre` suffix will be used for development purposes.
The commit modifies the version of `keep-ecdsa` package and updates the
versions of `keep-network/...` dependencies.