Use npm-version-bump and upstream-builds-query actions for versioning
This commit utilises two new actions in order to maintain correct
versioning, which from now on will not need to be configured manually.

The `upstream-builds-query` action will be used to retrieve information
about the versions of the upstream projects required for contracts
migration. Files `package.json` and `package-lock.json` will be then
updated based on that information.

At later stage of the job the npm package version will be bumped up
thanks to `npm-version-bump` action, based on the version of package
previously uploaded to npm registry for given environment.