Add retry logic to the electrs handle implementations
We create a clean separation of concerns between the bitcoin handle and
the caller (client.go). The bitcoin handle is in charge of trying as
hard as possible to retrieve valid information, and bubble up an error
if it can't. The caller is in charge of handling those errors and
knowing the proper fallbacks if the handle fails.