Log transaction hash as %s instead of %x
The `Hash` type has dedicated function to convert the value to String.
It is converted to hex with `0x` prefix, which seems more consistent for
ethereum. In keep-common code genreation we generate the loggers for
transaction id with `Hash().Hex()` function which produces `0x` prefixed
values. With this change we will be consistent.