More logging for TBTCSystem address configuration validations
We enhance coverage of cases that cause warnings to be logged on
TBCSystem address configuration validations.

The address is expected to be configured under
`config.Ethereum.ContractAddresses`.

Currently if the address is configured in
 - both config.Ethereum.ContractAddresses.TBTCSystem and config.Extensions.TBTC.TBTCSystem
 - both config.Ethereum.ContractAddresses.TBTCSystem and config.SanctionedApplications
we won't log a warning. With this change we will log a warning always if
the address is configured under a deprecated path to let
the operator know about needed configuration file cleanup.

We added a check for config.Ethereum.ContractAddresses.TBTCSystem and
config.Extensions.TBTC.TBTCSystem match to panic on missmatch. We
already have similiar check for SanctionedApplications.