Upgraded Golang to 1.16
We upgraded Golang to the latest version 1.16.

For docker image we use alpine 3.12, while the latest available source
image is based on 3.13. We use older alpine version as it comes with
Node 12 which is required by our solidity sources. Alpine 3.13 comes
with Node 14 which doesn't work well with our solidity package yet.

We also renamed package `python` to `python3`, as `python` is no longer
available.